requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
python-multipart
Pillow
torch
torchvision
scipy
slowapi
python-jose[cryptography]
python-dotenv
requests
